% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_managment.R
\name{add_beds}
\alias{add_beds}
\title{Add BEDs to Sample Sheet}
\usage{
add_beds(SCAR_obj, peak_dir, peak_type = peak_type, stringent = stringent)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{peak_dir}{Peak directory for SEACR output}

\item{stringent}{was SEACR run with the 'stringent' flag?}
}
\description{
Add BEDs to Sample Sheet
}
