% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_managment.R
\name{add_bgs}
\alias{add_bgs}
\title{Add Bedgraphs to Sample Sheet}
\usage{
add_bgs(SCAR_obj, alignment_dir)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{alignment_dir}{Directory for files to be dumped to.}
}
\description{
Add Bedgraphs to Sample Sheet
}
