% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_managment.R
\name{add_bws}
\alias{add_bws}
\title{Add bigwigs to Sample Sheet}
\usage{
add_bws(SCAR_obj, alignment_dir, comp_op = NA)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{alignment_dir}{Directory of aligned reads.}

\item{comp_op}{Operation for bamCompare}
}
\description{
Add bigwigs to Sample Sheet
}
