% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_managment.r
\name{add_cov}
\alias{add_cov}
\title{Add coverage tracks to Sample Sheet}
\usage{
add_cov(SCAR_obj, alignment_dir, comp_op = "comp")
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{alignment_dir}{Directory of aligned reads.}

\item{comp_op}{Operation for bamCompare}
}
\description{
Add coverage tracks to Sample Sheet
}
