% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_managment.R
\name{add_matrix}
\alias{add_matrix}
\title{Add matrix to Sample Sheet}
\usage{
add_matrix(SCAR_obj, plot_dir = plot_dir)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{plot_dir}{directory for matrix output}
}
\description{
Add matrix to Sample Sheet
}
