% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovbed_management.R
\name{add_ovbed}
\alias{add_ovbed}
\title{Add overlap beds to Sample Sheet}
\usage{
add_ovbed(SCAR_obj, alignment_dir)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{alignment_dir}{Directory for files to be dumped to.}
}
\description{
Add overlap beds to Sample Sheet
}
