% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_alignment.R
\name{bowtie2_align}
\alias{bowtie2_align}
\title{Bowtie2 Alignment}
\usage{
bowtie2_align(
  SCAR_obj,
  outdir = getwd(),
  alignment_mode = "end-to-end",
  min_fragment = NA,
  max_fragment = NA,
  mapq_val = 10,
  max_memory = "1G"
)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory for aligned reads.}

\item{alignment_mode}{Either 'end-to-end' or 'local'.}

\item{min_fragment}{Minimum fragment length (paired end).}

\item{max_fragment}{Maximum fragment length (paired end).}

\item{mapq_val}{Minimum MAPQ value to accept reads. Used to filter via
samtools post alignment}

\item{max_memory}{Maximum memory per thread for samtools.}
}
\description{
Bowtie2 Alignment
}
