% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_alignment.R
\name{bowtie2_index}
\alias{bowtie2_index}
\title{Bowtie2 Index Generation}
\usage{
bowtie2_index(
  SCAR_obj,
  genome_assembly,
  outdir = getwd(),
  index_name = "bowtie2_index"
)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{genome_assembly}{Path to genome fasta file.}

\item{outdir}{Output directory for genome index.}

\item{index_name}{The naming structure given to the index.}
}
\description{
Bowtie2 Index Generation
}
