% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolism.R
\name{calculate_score}
\alias{calculate_score}
\title{FUNCTION_TITLE}
\usage{
calculate_score(
  obj,
  assay,
  signatures = "KEGG",
  costume_label = "costume",
  ncores = 1,
  method_use = "AUC",
  ...
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{assay}{PARAM_DESCRIPTION}

\item{signatures}{PARAM_DESCRIPTION, Default: 'KEGG'}

\item{costume_label}{PARAM_DESCRIPTION, Default: 'costume'}

\item{ncores}{PARAM_DESCRIPTION, Default: 1}

\item{method_use}{PARAM_DESCRIPTION, Default: 'AUC'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[checkmate]{checkChoice}}, \code{\link[checkmate]{checkClass}}
 \code{\link[GSEABase]{import/export}}
 \code{\link[tools]{fileutils}}
 \code{\link[AUCell]{AUCell_buildRankings}}, \code{\link[AUCell]{AUCell_calcAUC}}, \code{\link[AUCell]{aucellResults-class}}
 \code{\link[glue]{glue}}
 \code{\link[Matrix]{colSums}}
 \code{\link[VISION]{c("Vision", "VISION", "VISION")}}, \code{\link[VISION]{analyze,Vision-method}}
}
