% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks_macs.R
\name{call_peaks_macs}
\alias{call_peaks_macs}
\title{MACS peak calling}
\usage{
call_peaks_macs(
  SCAR_obj,
  outdir = getwd(),
  process = "callpeak",
  in_form = NA,
  genome_size = NA,
  tag_size = NA,
  width = NA,
  mfold = NA,
  q_val = NA,
  p_val = NA,
  broad = FALSE
)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory.}

\item{process}{Name of the MACS function you want, def. is 'callpeaks'}

\item{in_form}{Format of input, default is to auto-determine}

\item{genome_size}{Size of mappable genome}

\item{tag_size}{size of adaptors, default is to auto-determine}

\item{width}{Width to use for model building scan, optional}

\item{mfold}{upper and lower limit for model building, optional}

\item{q_val}{min q value/FDR cutoff, optional}

\item{p_val}{Set p val instead of q, optional}

\item{broad}{Peaks are expected to be broad, TRUE or FALSE, default=FALSE}
}
\description{
MACS peak calling
}
