% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_peaks_SEACR.r
\name{call_peaks_SEACR}
\alias{call_peaks_SEACR}
\title{Peak Calling}
\usage{
call_peaks_SEACR(
  SCAR_obj,
  outdir = getwd(),
  num_thresh = NA,
  norm = NA,
  stringent = NA,
  sep = ""
)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory for peak files.}

\item{num_thresh}{P-value thresholding for peaks.}

\item{norm}{TRUE or FALSE - inputs are normalized or not}

\item{stringent}{TRUE or FALSE - stringent or relaxed analysis}

\item{sep}{separator for entries, default here is ""}
}
\description{
Peak Calling
}
