% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cov.R
\name{compare_bws}
\alias{compare_bws}
\title{Compare two bigWig input files}
\usage{
compare_bws(
  SCAR_obj,
  outdir = getwd(),
  comp_op = NA,
  bin_size = 1,
  skip_zeros = TRUE,
  skip_non_covered = FALSE,
  scale_factors = NA,
  out_type = NA,
  roi = NA
)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory.'}

\item{comp_op}{Operation for bamCompare}

\item{bin_size}{Bin size for coverage summary.}

\item{skip_zeros}{Skip bins where both files have 0 coverage, TRUE/FALSE}

\item{skip_non_covered}{Should regions without coverage be skipped.}

\item{scale_factors}{Takes a named vector, with the name being the
sample name, and the value being the scale factor.}

\item{out_type}{Output file format, bedgraph or bigwig}

\item{roi}{A specific region of interest to limit the process to}
}
\description{
Compare two bigWig input files
}
