% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake.R
\name{init_smk_script}
\alias{init_smk_script}
\title{FUNCTION_TITLE}
\usage{
init_smk_script(dir, config_list = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{dir}{PARAM_DESCRIPTION}

\item{config_list}{PARAM_DESCRIPTION, Default: NULL}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[checkmate]{checkPathForOutput}}, \code{\link[checkmate]{checkList}}
 \code{\link[yaml]{write_yaml}}
}
