% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_data}
\alias{load_data}
\title{Title ST data preprocess}
\usage{
load_data(
  data.dir,
  filename = "filtered_feature_bc_matrix.h5",
  slice = "slice1",
  assay_1 = "hg",
  assay_2 = NULL,
  assay_2_dir = NULL,
  ...
)
}
\arguments{
\item{data.dir}{Directory containing the H5 file specified by filename and the image data in a subdirectory called spatial}

\item{filename}{Name of H5 file containing the feature barcode matrix}

\item{slice}{Name for the stored image of the tissue slice}

\item{assay_1}{Name of the first assay}

\item{assay_2}{Name of the second assay}

\item{assay_2_dir}{Directory containing the 10X market format expression matrix}

\item{...}{PARAM_DESCRIPTION}
}
\value{
A Seurat object
}
\description{
read 10X ST raw data and quality control
}
\details{
DETAILS
}
\examples{
\dontrun{
dat <-
load_data(
data.dir = '/mnt/raid61/TNP_spatial/analysis/alignment/2021-1014-split/data/hg38/HF1/HF1/outs',
slice = 'HF1',
assay_1 = 'hg',
assay_2 = 'mm',
assay_2_dir = '/mnt/raid61/TNP_spatial/analysis/alignment/2021-1014-split/data/mm10/HF1/HF1/outs/filtered_feature_bc_matrix/'
)
}
}
\seealso{
\code{\link[Seurat]{Load10X_Spatial}}, \code{\link[Seurat]{Read10X}}, \code{\link[Seurat]{reexports}}
 \code{\link[checkmate]{checkTRUE}}
 \code{\link[glue]{glue}}
}
