% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bgs.R
\name{make_bgs}
\alias{make_bgs}
\title{Make bedgraphs}
\usage{
make_bgs(SCAR_obj, outdir = getwd(), pair_lr = FALSE, frag_size = FALSE)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory.}

\item{pair_lr}{TRUE or FALSE, calculate coverage as number of
frags covering each bp (paired)}

\item{frag_size}{Use stated fragment size from pairs instead
of read length (paired)}
}
\description{
Make bedgraphs
}
