% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix.R
\name{make_matrix}
\alias{make_matrix}
\title{Convert coverage file to matrix}
\usage{
make_matrix(
  SCAR_obj,
  outdir = getwd(),
  primary = "reference-point",
  s_n_c = NA,
  in_str = NA,
  regions = NA
)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory.'}

\item{primary}{Main command; either 'reference-point' or 'scale-regions'}

\item{s_n_c}{If desired, select a stored bigWig set for processing;
"s" for sample, "n" for control, "c" for compare}

\item{in_str}{A list of input strings to be converted to a bash command for
computeMatrix}

\item{regions}{A bed file defining regions to be plotted; if none entered,
the peak file from peak finding will be used}
}
\description{
Convert coverage file to matrix
}
