% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tracks_opts.R
\name{make_tracks_opts}
\alias{make_tracks_opts}
\title{Generate coverage tracks}
\usage{
make_tracks_opts(
  SCAR_obj,
  outdir = getwd(),
  compare = FALSE,
  scale_method = NA,
  comp_op = "log2",
  bin_size = 1,
  normalize_using = NA,
  genome_size = NA,
  skip_non_covered = NA,
  min_fragment = NA,
  max_fragment = NA,
  extend_reads = NA,
  scale_factors = NA,
  split_strands = NA,
  library_type = NA,
  center_reads = NA,
  out_type = NA,
  temp_dir = "./temp"
)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory.}

\item{compare}{TRUE or FALSE - should bamCompare or bamCoverage be performed}

\item{scale_method}{Default is 'readCount', other options are 'SES' and 'NONE'}

\item{comp_op}{Operation for bamCompare}

\item{bin_size}{Bin size for coverage summary.}

\item{normalize_using}{Either 'CPM' or 'RPGC'/'RPKM'.}

\item{genome_size}{Effective genome size, req'd for RPGC/RPKM.}

\item{skip_non_covered}{Should regions without coverage be skipped.}

\item{min_fragment}{Minimum fragment length.}

\item{max_fragment}{Maximum fragment length.}

\item{extend_reads}{Distance to extend single-end reads.
Set to NA to not extend reads.}

\item{scale_factors}{Takes a named vector, with the name being the
sample name, and the value being the scale factor.
If set will override 'normalize_using' option.}

\item{split_strands}{For RNA-seq, whether to split the strands into
positive and minus strand files.}

\item{library_type}{If split_strands is TRUE, specify library chemistry
as either 'dUTP' or 'ligation'.}

\item{center_reads}{Should reads be centered based on fragment length? Helps with signal around enriched regions}

\item{out_type}{Output file format, bedgraph or bigwig}

\item{temp_dir}{Temporary directory to write files to.}
}
\description{
Generate coverage tracks
}
