% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_analysis.R
\name{pf_analysis}
\alias{pf_analysis}
\title{Protected fragment analysis}
\usage{
pf_analysis(SCAR_obj, outdir = getwd(), bed_file = bed_file)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory.}

\item{bed_file}{A bed file of regions to be checked for local low signal sites in aligned bams}
}
\description{
Protected fragment analysis
}
