% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot matrix output in profile}
\usage{
plot_profile(
  SCAR_obj,
  outdir = getwd(),
  matrix = NA,
  plot_name = NA,
  plot_opts = NA,
  clust = NA
)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory.}

\item{matrix}{If desired, provide a matrix for processing; by default,
the stored matrix is used}

\item{plot_name}{name for plot with extension (.pdf or .png)}

\item{plot_opts}{A list of input strings to be converted to a bash command for
the plot function}

\item{clust}{If given, data will be clustered into the number of groups
equal to the value entered (integers only) via hclust}
}
\description{
Plot matrix output in profile
}
