% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_handler.R
\name{pull_setting}
\alias{pull_setting}
\title{Pull Settings}
\usage{
pull_setting(SCAR_obj, setting)
}
\arguments{
\item{SCAR_obj}{SCAR object, holds data for use in all the functions.}

\item{setting}{Setting to pull.}
}
\description{
Pull Settings
}
