% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_reads}
\alias{retrieve_reads}
\title{Retrieve Reads from SRA}
\usage{
retrieve_reads(SCAR_obj, outdir = getwd())
}
\arguments{
\item{SCAR_obj}{SCAR object, holds data for use in all the functions..}

\item{outdir}{Directory to download files to.}
}
\description{
Retrieve Reads from SRA
}
