% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{run_CARD}
\alias{run_CARD}
\title{FUNCTION_TITLE}
\usage{
run_CARD(
  obj,
  reference,
  sc_meta,
  assay = "Spatial",
  minCountGene = 0,
  minCountSpot = 0,
  ...
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{reference}{PARAM_DESCRIPTION}

\item{sc_meta}{PARAM_DESCRIPTION}

\item{assay}{PARAM_DESCRIPTION, Default: 'Spatial'}

\item{minCountGene}{PARAM_DESCRIPTION, Default: 0}

\item{minCountSpot}{PARAM_DESCRIPTION, Default: 0}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[checkmate]{checkClass}}
 \code{\link[Seurat]{reexports}}
 \code{\link[assertable]{assert_colnames}}
 \code{\link[CARD]{createCARDObject}}, \code{\link[CARD]{CARD_deconvolution}}
}
