% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{run_cooccur}
\alias{run_cooccur}
\title{FUNCTION_TITLE}
\usage{
run_cooccur(
  obj,
  lr_network,
  idents = NULL,
  spots = NULL,
  threshold_gene_exp = 0,
  slot_1 = "counts",
  assay_1 = "Spatial",
  slot_2 = "counts",
  assay_2 = "Spatial",
  prob = "comb",
  cores = 4,
  ...
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{lr_network}{PARAM_DESCRIPTION}

\item{idents}{PARAM_DESCRIPTION, Default: NULL}

\item{spots}{PARAM_DESCRIPTION, Default: NULL}

\item{threshold_gene_exp}{PARAM_DESCRIPTION, Default: 0}

\item{slot_1}{PARAM_DESCRIPTION, Default: 'counts'}

\item{assay_1}{PARAM_DESCRIPTION, Default: 'Spatial'}

\item{slot_2}{PARAM_DESCRIPTION, Default: 'counts'}

\item{assay_2}{PARAM_DESCRIPTION, Default: 'Spatial'}

\item{prob}{PARAM_DESCRIPTION, Default: 'comb'}

\item{cores}{PARAM_DESCRIPTION, Default: 4}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[checkmate]{checkFALSE}}
 \code{\link[Seurat]{reexports}}
 \code{\link[stringr]{case}}
 \code{\link[Matrix]{colSums}}
 \code{\link[tibble]{tibble}}
 \code{\link[parallel]{mclapply}}
 \code{\link[cooccur]{cooccur}}
 \code{\link[plyr]{mapvalues}}
}
