% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{run_DWLS}
\alias{run_DWLS}
\title{Perform de-convolution by DWLS}
\usage{
run_DWLS(
  obj,
  reference,
  assay = "Spatial",
  group.by = "cluster",
  cutoff = 2,
  n_cell = 50,
  ...
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{reference}{PARAM_DESCRIPTION}

\item{assay}{PARAM_DESCRIPTION, Default: 'Spatial'}

\item{group.by}{PARAM_DESCRIPTION, Default: 'cluster'}

\item{cutoff}{PARAM_DESCRIPTION, Default: 2}

\item{n_cell}{PARAM_DESCRIPTION, Default: 50}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Perform de-convolution by DWLS, more detailed information refer to \url{https://github.com/dtsoucas/DWLS}
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Matrix]{character(0)}}
 \code{\link[Seurat]{reexports}}
 \code{\link[Giotto]{enrich_deconvolution}}, \code{\link[Giotto]{spot_deconvolution}}
}
