% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{run_iascore}
\alias{run_iascore}
\title{Generate a interaction score for each ligant-receptor pair.}
\usage{
run_iascore(
  obj,
  lr_network,
  threshold_gene_exp = 0,
  slot_1 = "data",
  assay_1 = "Spatial",
  slot_2 = "data",
  assay_2 = "Spatial",
  cores = 4,
  ...
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{lr_network}{PARAM_DESCRIPTION}

\item{threshold_gene_exp}{PARAM_DESCRIPTION, Default: 0}

\item{slot_1}{PARAM_DESCRIPTION, Default: 'data'}

\item{assay_1}{PARAM_DESCRIPTION, Default: 'Spatial'}

\item{slot_2}{PARAM_DESCRIPTION, Default: 'data'}

\item{assay_2}{PARAM_DESCRIPTION, Default: 'Spatial'}

\item{cores}{PARAM_DESCRIPTION, Default: 4}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{case}}
 \code{\link[Seurat]{reexports}}
 \code{\link[Matrix]{colSums}}
 \code{\link[tibble]{tibble}}
 \code{\link[parallel]{mclapply}}
}
