% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{run_rctd}
\alias{run_rctd}
\title{FUNCTION_TITLE}
\usage{
run_rctd(
  obj,
  reference,
  doublet_mode = "full",
  assay = "Spatial",
  ncores = 1,
  ...
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{reference}{PARAM_DESCRIPTION}

\item{doublet_mode}{PARAM_DESCRIPTION, Default: 'full'}

\item{assay}{PARAM_DESCRIPTION, Default: 'Spatial'}

\item{ncores}{PARAM_DESCRIPTION, Default: 1}

\item{...}{PARAM_DESCRIPTION}
}
\value{
a data.frame, each column means a cell type and row names is the cell id.
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[spacexr]{SpatialRNA}}, \code{\link[spacexr]{create.RCTD}}
 \code{\link[Seurat]{reexports}}
 \code{\link[Matrix]{colSums}}
}
