\name{scar-internal}\title{Internal shape-constrained additive regression functions}
\alias{aggsum}
\description{
  Internal functions for the package \pkg{scar}
}
\details{
  These internal functions are not to be called by the user. 

  For a given real sequence \eqn{a_1,..., a_n}, the function \code{\link{aggsum}} 
  returns the partial sums 
  \eqn{a_1, a_1+a_2, a_1+a_2+a_3,..., \sum_{i=1}^n a_i}. 
}

\keyword{ internal }
