\name{SCAR-package}

\alias{SCAR-package}

\alias{SCAR}

\docType{package}

\title{
  \packageTitle{SCAR}
  }

\description{
  \packageDescription{SCAR}
  }

\details{
  The DESCRIPTION file:
  \packageDESCRIPTION{SCAR}
  \packageIndices{SCAR}
  }

\author{
  \packageAuthor{SCAR}
  Maintainer: \packageMaintainer{SCAR}
  }
  
\references{
  }
  
\keyword{ SCAR }

\seealso{
  }

\examples{
  }
