% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAR_obj.R
\name{SCAR_maker}
\alias{SCAR_maker}
\title{SCAR Object Constructor Function}
\usage{
SCAR_maker(
  analysis_type,
  sample_sheet,
  sep = "\\t",
  paired = NA,
  ncores = NA,
  compare = NA,
  comp_op = NA
)
}
\arguments{
\item{analysis_type}{One of 'ChIP-seq', 'ChEC-seq', or 'SChEC-seq'.}

\item{sample_sheet}{Either a data.frame or delimited file.}

\item{sep}{If the sample sheet is a file, this specifies the delimiter.}

\item{paired}{Whether the run is paired (TRUE) or unpaired (FALSE)}

\item{ncores}{The number of cores/threads to use.}

\item{compare}{TRUE/FALSE perform bamCompare instead of Coverage}

\item{comp_op}{Single operation for comparison of coverage}
}
\value{
A SCAR object.
}
\description{
SCAR Object Constructor Function
}
