% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCAR_obj.R
\docType{class}
\name{SCAR_obj-class}
\alias{SCAR_obj-class}
\title{SCAR Class}
\description{
SCAR Class
}
\section{Slots}{

\describe{
\item{\code{sample_sheet}}{Sample sheet.}

\item{\code{settings}}{Various run settings.}
}}

