% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SpatialInteractionPlot}
\alias{SpatialInteractionPlot}
\title{FUNCTION_TITLE}
\usage{
SpatialInteractionPlot(
  obj,
  idents = NULL,
  spots = NULL,
  assay_1 = "Spatial",
  gene_1 = NULL,
  assay_2 = "Spatial",
  gene_2 = NULL,
  slot = "data",
  pt.size.factor = 1,
  max.cutoff = NA,
  ...
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{idents}{PARAM_DESCRIPTION, Default: NULL}

\item{spots}{PARAM_DESCRIPTION, Default: NULL}

\item{assay_1}{PARAM_DESCRIPTION, Default: 'Spatial'}

\item{gene_1}{PARAM_DESCRIPTION, Default: NULL}

\item{assay_2}{PARAM_DESCRIPTION, Default: 'Spatial'}

\item{gene_2}{PARAM_DESCRIPTION, Default: NULL}

\item{slot}{PARAM_DESCRIPTION, Default: 'data'}

\item{pt.size.factor}{PARAM_DESCRIPTION, Default: 1}

\item{max.cutoff}{PARAM_DESCRIPTION, Default: NA}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[checkmate]{checkFALSE}}
 \code{\link[Seurat]{reexports}}, \code{\link[Seurat]{SpatialPlot}}
 \code{\link[glue]{glue}}
}
