% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_analysis.R
\name{pf_analysis}
\alias{pf_analysis}
\title{Protected fragment analysis}
\usage{
pf_analysis(
  SCAR_obj,
  outdir = getwd(),
  in_bam = NA,
  in_bg = NA,
  thresh = 0,
  bed_file = bed_file
)
}
\arguments{
\item{SCAR_obj}{SCAR object.}

\item{outdir}{Output directory.}

\item{in_bam}{path to an input bam, leave blank if none}

\item{in_bg}{path to an input bedgraph, leave blank if none}

\item{thresh}{a cutoff for 'low' signal. Coverage regions under this threshold will be compared to
the input bam file as 'low signal' regions}

\item{bed_file}{A bed file of regions to be checked for local low signal sites in aligned bams}
}
\description{
Protected fragment analysis
}
